requirements.txt
Flask==2.2.3
pdfminer.six==20201018
spacy==3.7.2
setuptools==80.9.0
numpy==2.1.3
Cython==3.1.2
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.3.0
langdetect==1.0.9
Werkzeug==2.2.3
